requirements.txt
astroid==1.5.3
beautifulsoup4==4.4.1
certifi==2017.4.17
chardet==3.0.3
colorama==0.3.9
cssutils==1.0.2
dnspython==1.15.0
enum34==1.1.6
future==0.16.0
idna==2.5
isort==4.2.14
lazy-object-proxy==1.3.1
lxml==3.8.0
mccabe==0.6.1
protobuf==3.0.0b2
pycaption==1.0.0
pycryptodomex==3.4.6
pylint==1.7.1
pysubs2==0.2.1
requests==2.17.3
six==1.10.0
tqdm==4.14.0
urllib3==1.21.1
wrapt==1.10.10
xmltodict==0.11.0
